SAA-1150 common component to be used by event handlers for deallocations.
